src/main/java/org/parsky/sequence/MatchedTextSequenceMatcher.java
package org.parsky.sequence;

import org.parsky.sequence.model.SequenceMatcherRequest;
import org.parsky.sequence.model.SequenceMatcherResult;

public class MatchedTextSequenceMatcher implements SequenceMatcher {
    private final SequenceMatcher sequenceMatcher;

    public MatchedTextSequenceMatcher(SequenceMatcher sequenceMatcher) {
        this.sequenceMatcher = sequenceMatcher;
    }

    @Override
    public SequenceMatcherResult matches(SequenceMatcherRequest sequenceMatcherRequest) {
        SequenceMatcherResult result = sequenceMatcher.matches(sequenceMatcherRequest);

        if (result.matched()) {
            return SequenceMatcherResult.match(result.getJump(), sequenceMatcherRequest.text(result.getJump()));
        } else {
            return result;
        }
    }
}
